+++ server.c
@@ -33,7 +33,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 #define EXEC_NARGS 20
 char *exec_args[EXEC_NARGS + 1];
-int exec_argc = 0;
+unsigned exec_argc = 0;
 
 int servalShutdown = 0;
 
@@ -74,7 +74,7 @@ int server_pid()
 void server_save_argv(int argc, const char *const *argv)
 {
     /* Save our argv[] to use for relaunching */
-    for (exec_argc = 0; exec_argc < argc && exec_argc < EXEC_NARGS; ++exec_argc)
+    for (exec_argc = 0; exec_argc < (unsigned)argc && exec_argc < EXEC_NARGS; ++exec_argc)
       exec_args[exec_argc] = strdup(argv[exec_argc]);
     exec_args[exec_argc] = NULL;
 }
@@ -392,7 +392,7 @@ void crash_handler(int signal)
   
   BACKTRACE;
   if (config.server.respawn_on_crash) {
-    int i;
+    unsigned i;
     for(i=0;i<overlay_interface_count;i++)
       if (overlay_interfaces[i].alarm.poll.fd>-1)
 	close(overlay_interfaces[i].alarm.poll.fd);
